+++ lib/src/utils/github_snitch_instance.dart
@@ -23,6 +23,7 @@ class GhSnitchInstance {
   String? token;
   String? owner;
   String? repo;
+  int? maxDuplicatedReports;
   OnReport? onReport;
   late GhRequest ghRequest;
   static GhSnitchInstance get instance => GhSnitchInstance();
@@ -141,10 +142,12 @@ class GhSnitchInstance {
       {required String token,
       required String owner,
       required String repo,
+      required int maxDuplicatedReports,
       OnReport onReport}) {
     this.token = token;
     this.owner = owner;
     this.repo = repo;
+    this.maxDuplicatedReports = maxDuplicatedReports;
     this.onReport = onReport;
     if (token.isEmpty || owner.isEmpty || repo.isEmpty) {
       log("🔴 Echec to initialize GhSnitch");
@@ -214,7 +217,7 @@ class GhSnitchInstance {
   /// Checks if an issue with a similar `body` has already been reported in the repository.
   /// It takes the `body` as required parametes.
   Future<bool> isAlreadyReported(String body) async {
-    bool isAlreadyReported = false;
+    bool isDuplicated = false;
     body = body
         .replaceAll("```", "")
         .substring(0, math.min(body.length, 255))
@@ -223,13 +226,17 @@ class GhSnitchInstance {
         "https://api.github.com/search/issues?q=repo:$owner/$repo+is:issue+is:open+$body";
     GhResponse ghResponse = await ghRequest.request("GET", url, isSearch: true);
     if (ghResponse.statusCode == 200) {
-      isAlreadyReported = ghResponse.response['total_count'] != 0;
-      if (isAlreadyReported) {
-        await submitComment(
-            ghResponse.response['items'][0][issueNumber].toString(), "+1");
+      final count = ghResponse.response['total_count'];
+      isDuplicated = count != 0;
+      if (isDuplicated) {
+        final comments = ghResponse.response["items"].first["comments"];
+        if (comments < maxDuplicatedReports) {
+          await submitComment(
+              ghResponse.response['items'][0][issueNumber].toString(), "+1");
+        }
       }
     }
-    return isAlreadyReported;
+    return isDuplicated;
   }
 
   /// Retrieves all the issues in the repository that contain the specified `userId`.